# openapi.yaml
openapi: 3.1.0
info:
  title: Collab Canvas Aid API
  version: 0.1.0
  description: |
    HTTP and WebSocket contract expected by the frontend in frontend/.

    The checked-in frontend currently uses an in-browser MockBackend rather than
    an HTTP client. This specification maps the Services interface and the API
    surface described by frontend/README.md to a backend contract. Authentication
    is delegated to an external identity provider; the API receives its resulting
    bearer token.
servers:
  - url: https://api.example.com
    description: Replace with the deployed API origin.
...
paths:
...
  /v1/sessions/{id}/realtime:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      tags: [Collaboration]
      operationId: connectRealtime
      summary: Upgrade to the session collaboration WebSocket
      description: |
        This GET is an HTTP Upgrade request to WebSocket. The bearer token may
        be the signed-in user's token or the short-lived collaboration token
        returned by the guest join flow. The message contract is described by
        x-websocket below because OpenAPI does not standardize WebSocket frames.
      security:
        - bearerAuth: []
      responses:
        '101':
          description: Switching Protocols; WebSocket connection established.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      x-websocket:
        protocolVersion: 1
        clientToServer:
          - $ref: '#/components/schemas/JoinRoomMessage'
          - $ref: '#/components/schemas/DocumentUpdateMessage'
          - $ref: '#/components/schemas/PresenceUpdateMessage'
          - $ref: '#/components/schemas/PingMessage'
        serverToClient:
          - $ref: '#/components/schemas/RoomJoinedMessage'
          - $ref: '#/components/schemas/DocumentUpdateMessage'
          - $ref: '#/components/schemas/PresenceSnapshotMessage'
          - $ref: '#/components/schemas/PresenceUpdateMessage'
          - $ref: '#/components/schemas/PermissionChangedMessage'
          - $ref: '#/components/schemas/SessionEndedMessage'
          - $ref: '#/components/schemas/RealtimeErrorMessage'
          - $ref: '#/components/schemas/PongMessage'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Signed-in user access token, or a short-lived participant collaboration
        credential returned after a guest joins. Guest-link URLs are only used
        for the initial public resolve/join exchange.
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      description: Interview session identifier.
      schema:
        type: string
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated principal lacks permission for this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource does not exist or is not visible to the caller.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    SessionState:
      type: string
      enum: [draft, live, ended, archived]
    Role:
      type: string
      enum: [owner, interviewer, candidate, observer]
    InterviewSession:
      type: object
      required:
        - id
        - ownerUserId
        - title
        - prompt
        - state
        - candidateEditingEnabled
        - durationMinutes
        - scheduledAt
        - startedAt
        - endedAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        ownerUserId:
          type: string
        title:
          type: string
        prompt:
          type: string
        state:
          $ref: '#/components/schemas/SessionState'
        candidateEditingEnabled:
          type: boolean
        durationMinutes:
          type: integer
          minimum: 1
        scheduledAt:
          oneOf:
            - $ref: '#/components/schemas/EpochMillis'
            - type: 'null'
        startedAt:
          oneOf:
            - $ref: '#/components/schemas/EpochMillis'
            - type: 'null'
        endedAt:
          oneOf:
            - $ref: '#/components/schemas/EpochMillis'
            - type: 'null'
        createdAt:
          $ref: '#/components/schemas/EpochMillis'
        updatedAt:
          $ref: '#/components/schemas/EpochMillis'
...
    Participant:
      type: object
      required: [id, sessionId, userId, displayName, role, color, joinedAt, leftAt, online]
      properties:
        id:
          type: string
        sessionId:
          type: string
        userId:
          oneOf:
            - type: string
            - type: 'null'
        displayName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        color:
          type: string
        joinedAt:
          $ref: '#/components/schemas/EpochMillis'
        leftAt:
          oneOf:
            - $ref: '#/components/schemas/EpochMillis'
            - type: 'null'
        online:
          type: boolean
    Presence:
      type: object
      required: [participantId, displayName, color, cursor, selection]
      properties:
        participantId:
          type: string
        displayName:
          type: string
        color:
          type: string
        cursor:
          oneOf:
            - $ref: '#/components/schemas/Point'
            - type: 'null'
        selection:
          type: array
          items:
            type: string
...
    CanvasDoc:
      type: object
      required: [schemaVersion, elements, order]
      properties:
        schemaVersion:
          type: integer
          minimum: 1
        elements:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/CanvasElement'
        order:
          type: array
          description: Element IDs in draw order, from bottom to top.
          items:
            type: string
    CanvasElement:
      oneOf:
        - $ref: '#/components/schemas/NodeElement'
        - $ref: '#/components/schemas/StickyElement'
        - $ref: '#/components/schemas/TextElement'
        - $ref: '#/components/schemas/ConnectorElement'
        - $ref: '#/components/schemas/StrokeElement'
      discriminator:
        propertyName: kind
        mapping:
          node: '#/components/schemas/NodeElement'
          sticky: '#/components/schemas/StickyElement'
          text: '#/components/schemas/TextElement'
          connector: '#/components/schemas/ConnectorElement'
          stroke: '#/components/schemas/StrokeElement'
    ElementBase:
      type: object
      required: [id, createdBy, createdAt, updatedAt]
      properties:
        id:
          type: string
        createdBy:
          type: string
        createdAt:
          $ref: '#/components/schemas/EpochMillis'
        updatedAt:
          $ref: '#/components/schemas/EpochMillis'
    ElementColor:
      type: string
      enum: [neutral, accent, amber, rose]
    NodeElement:
      allOf:
        - $ref: '#/components/schemas/ElementBase'
        - type: object
          required: [kind, componentType, x, y, width, height, label, color]
          properties:
            kind:
              const: node
            componentType:
              type: string
            x:
              type: number
            y:
              type: number
            width:
              type: number
            height:
              type: number
            label:
              type: string
            description:
              type: string
            color:
              $ref: '#/components/schemas/ElementColor'
    StickyElement:
      allOf:
        - $ref: '#/components/schemas/ElementBase'
        - type: object
          required: [kind, x, y, width, height, text, color]
          properties:
            kind:
              const: sticky
            x:
              type: number
            y:
              type: number
            width:
              type: number
            height:
              type: number
            text:
              type: string
            color:
              $ref: '#/components/schemas/ElementColor'
    TextElement:
      allOf:
        - $ref: '#/components/schemas/ElementBase'
        - type: object
          required: [kind, x, y, text, size, color]
          properties:
            kind:
              const: text
            x:
              type: number
            y:
              type: number
            text:
              type: string
            size:
              type: number
            color:
              $ref: '#/components/schemas/ElementColor'
    ConnectorEndpoint:
      oneOf:
        - type: object
          required: [elementId]
          properties:
            elementId:
              type: string
        - $ref: '#/components/schemas/Point'
    ConnectorElement:
      allOf:
        - $ref: '#/components/schemas/ElementBase'
        - type: object
          required: [kind, from, to, style, arrowStart, arrowEnd, dashed, width, color]
          properties:
            kind:
              const: connector
            from:
              $ref: '#/components/schemas/ConnectorEndpoint'
            to:
              $ref: '#/components/schemas/ConnectorEndpoint'
            style:
              type: string
              enum: [straight, elbow, curved]
            arrowStart:
              type: boolean
            arrowEnd:
              type: boolean
            dashed:
              type: boolean
            width:
              type: number
            label:
              type: string
            color:
              $ref: '#/components/schemas/ElementColor'
    StrokeElement:
      allOf:
        - $ref: '#/components/schemas/ElementBase'
        - type: object
          required: [kind, points, color, width, tool]
          properties:
            kind:
              const: stroke
            points:
              type: array
              items:
                type: array
                prefixItems:
                  - type: number
                  - type: number
                minItems: 2
                maxItems: 2
            color:
              $ref: '#/components/schemas/ElementColor'
            width:
              type: number
            tool:
              type: string
              enum: [pen, highlighter]
    CanvasElementPatch:
      type: object
      description: Partial element object applied by an update operation.
      additionalProperties: true
    CanvasOperation:
      oneOf:
        - $ref: '#/components/schemas/AddOperation'
        - $ref: '#/components/schemas/UpdateOperation'
        - $ref: '#/components/schemas/DeleteOperation'
        - $ref: '#/components/schemas/ReorderOperation'
        - $ref: '#/components/schemas/ClearOperation'
    AddOperation:
      type: object
      required: [type, element]
      properties:
        type:
          const: add
        element:
          $ref: '#/components/schemas/CanvasElement'
    UpdateOperation:
      type: object
      required: [type, id, patch, at]
      properties:
        type:
          const: update
        id:
          type: string
        patch:
          $ref: '#/components/schemas/CanvasElementPatch'
        at:
          $ref: '#/components/schemas/EpochMillis'
    DeleteOperation:
      type: object
      required: [type, ids]
      properties:
        type:
          const: delete
        ids:
          type: array
          minItems: 1
          items:
            type: string
    ReorderOperation:
      type: object
      required: [type, id, to]
      properties:
        type:
          const: reorder
        id:
          type: string
        to:
          type: string
          enum: [front, back]
    ClearOperation:
      type: object
      required: [type]
      properties:
        type:
          const: clear
    CanvasEnvelope:
      type: object
      required: [clientOperationId, actorId, sentAt, operation]
      properties:
        clientOperationId:
          type: string
          description: Unique per client operation; used for deduplication.
        actorId:
          type: string
        sentAt:
          $ref: '#/components/schemas/EpochMillis'
        operation:
          $ref: '#/components/schemas/CanvasOperation'
...
    ErrorCode:
      type: string
      enum: [not_found, revoked, expired, at_capacity, forbidden, ended, invalid]
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    Point:
      type: object
      required: [x, y]
      properties:
        x:
          type: number
        y:
          type: number
    EpochMillis:
      type: integer
      format: int64
      description: Unix epoch timestamp in milliseconds.
    RealtimeEnvelope:
      type: object
      required: [type, version, sessionId]
      properties:
        type:
          type: string
        version:
          type: integer
          const: 1
        sessionId:
          type: string
        correlationId:
          type: string
    JoinRoomMessage:
      allOf:
        - $ref: '#/components/schemas/RealtimeEnvelope'
        - type: object
          required: [type, participantId]
          properties:
            type:
              const: join_room
            participantId:
              type: string
    DocumentUpdateMessage:
      allOf:
        - $ref: '#/components/schemas/RealtimeEnvelope'
        - type: object
          required: [type, envelopes]
          properties:
            type:
              const: document_update
            envelopes:
              type: array
              minItems: 1
              items:
                $ref: '#/components/schemas/CanvasEnvelope'
    PresenceUpdateMessage:
      allOf:
        - $ref: '#/components/schemas/RealtimeEnvelope'
        - type: object
          required: [type, presence]
          properties:
            type:
              const: presence_update
            presence:
              $ref: '#/components/schemas/Presence'
    PingMessage:
      allOf:
        - $ref: '#/components/schemas/RealtimeEnvelope'
        - type: object
          required: [type]
          properties:
            type:
              const: ping
    RoomJoinedMessage:
      allOf:
        - $ref: '#/components/schemas/RealtimeEnvelope'
        - type: object
          required: [type, snapshot, participants, presences, cursor]
          properties:
            type:
              const: room_joined
            snapshot:
              $ref: '#/components/schemas/CanvasDoc'
            participants:
              type: array
              items:
                $ref: '#/components/schemas/Participant'
            presences:
              type: array
              items:
                $ref: '#/components/schemas/Presence'
            cursor:
              type: integer
              minimum: 0
    PresenceSnapshotMessage:
      allOf:
        - $ref: '#/components/schemas/RealtimeEnvelope'
        - type: object
          required: [type, presences]
          properties:
            type:
              const: presence_snapshot
            presences:
              type: array
              items:
                $ref: '#/components/schemas/Presence'
    PermissionChangedMessage:
      allOf:
        - $ref: '#/components/schemas/RealtimeEnvelope'
        - type: object
          required: [type, session]
          properties:
            type:
              const: permission_changed
            session:
              $ref: '#/components/schemas/InterviewSession'
    SessionEndedMessage:
      allOf:
        - $ref: '#/components/schemas/RealtimeEnvelope'
        - type: object
          required: [type, session]
          properties:
            type:
              const: session_ended
            session:
              $ref: '#/components/schemas/InterviewSession'
    RealtimeErrorMessage:
      allOf:
        - $ref: '#/components/schemas/RealtimeEnvelope'
        - type: object
          required: [type, error]
          properties:
            type:
              const: error
            error:
              $ref: '#/components/schemas/ErrorResponse'
    PongMessage:
      allOf:
        - $ref: '#/components/schemas/RealtimeEnvelope'
        - type: object
          required: [type]
          properties:
            type:
              const: pong